@@ -15,7 +15,8 @@
   "author": "dsshard",
   "scripts": {
     "start": "NODE_ENV=development && webpack-cli --watch --mode development",
-    "build": "NODE_ENV=production && webpack-cli --mode production"
+    "build": "NODE_ENV=production && webpack-cli --mode production",
+    "test": "jest ."
   },
   "license": "MIT",
   "peerDependencies": {
@@ -23,22 +24,26 @@
     "prop-types": ">=15.5.x"
   },
   "dependencies": {
-    "react": ">=16.3.x",
-    "prop-types": ">=15.5.x"
+    "prop-types": ">=15.5.x",
+    "react": ">=16.3.x"
   },
   "devDependencies": {
-    "@babel/core": "7.5.5",
-    "@babel/plugin-proposal-class-properties": "7.5.5",
-    "@babel/preset-env": "7.5.5",
-    "@babel/preset-react": "7.0.0",
-    "babel-eslint": "10.0.2",
-    "babel-loader": "8.0.6",
-    "eslint": "6.1.0",
-    "eslint-config-airbnb": "18.0.1",
-    "eslint-plugin-import": "2.18.2",
-    "eslint-plugin-jsx-a11y": "6.2.3",
-    "eslint-plugin-react": "7.14.3",
-    "webpack": "4.39.2",
-    "webpack-cli": "3.3.6"
+    "@babel/core": "7.10.5",
+    "@babel/plugin-proposal-class-properties": "7.10.4",
+    "@babel/preset-env": "7.10.4",
+    "@babel/preset-react": "7.10.4",
+    "babel-eslint": "10.1.0",
+    "babel-loader": "8.1.0",
+    "babel-plugin-transform-react-remove-prop-types": "0.4.24",
+    "eslint": "7.5.0",
+    "eslint-config-airbnb": "18.2.0",
+    "eslint-plugin-import": "2.22.0",
+    "eslint-plugin-jsx-a11y": "6.3.1",
+    "eslint-plugin-react": "7.20.3",
+    "jest": "26.1.0",
+    "react-dom": "16.8.6",
+    "react-test-renderer": "16.8.6",
+    "webpack": "4.43.0",
+    "webpack-cli": "3.3.12"
   }
 }
