@@ -1,9 +1,10 @@
-const path = require('path');
-
 module.exports = {
   parser: 'babel-eslint',
   env: {
     browser: true,
   },
-  extends: 'airbnb'
+  rules: {
+    'react/destructuring-assignment': [0],
+  },
+  extends: 'airbnb',
 };
