@@ -1,5 +1,4 @@
 /* eslint react/sort-comp: [0], camelcase: [0] */
-
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import Context from './context';
@@ -21,18 +20,15 @@ class ValidatorWrapper extends Component {
 
   unregisterField = (field) => {
     const index = this.fields.indexOf(field);
-    if (field && index > -1) {
-      this.fields.splice(index, 1);
-    }
+    if (index > -1) this.fields.splice(index, 1);
   }
 
   getField = (id) => this.fields.find((field) => field.props.id === id) || null
 
   validate = () => {
-    const { stopAtFirstError } = this.props;
     let prevResult;
     const statuses = this.fields.map((field) => {
-      if (stopAtFirstError && prevResult && prevResult.isValid === false) {
+      if (this.props.stopAtFirstError && prevResult && prevResult.isValid === false) {
         return null;
       }
       prevResult = field.validate();
@@ -40,8 +36,8 @@ class ValidatorWrapper extends Component {
     });
 
     const errors = statuses
-      .filter((instance) => instance)
-      .filter((instance) => instance.isValid === false);
+      .filter((inst) => inst)
+      .filter((inst) => inst.isValid === false);
 
     if (errors.length) {
       return Object.assign(errors[0], {
@@ -52,16 +48,15 @@ class ValidatorWrapper extends Component {
   }
 
   render() {
-    const { registerField, unregisterField } = this;
-    const { children } = this.props;
     return (
-      <Context.Provider value={{ registerField, unregisterField }}>
-        {children}
+      <Context.Provider
+        value={{ registerField: this.registerField, unregisterField: this.unregisterField }}
+      >
+        {this.props.children}
       </Context.Provider>
     );
   }
 }
-
 
 ValidatorWrapper.propTypes = {
   children: PropTypes.node.isRequired,
